Core & Internals: fixed more typos and rebase problems; Fix #2315

+++ lib/rucio/tests/test_bin_rucio.py
@@ -43,8 +43,6 @@ from rucio.client.ruleclient import RuleClient
 from rucio.common.config import config_get
 from rucio.common.types import InternalScope, InternalAccount
 from rucio.common.utils import generate_uuid, md5, render_json
-from rucio.core.account_counter import increase
-from rucio.core.rse import add_rse_attribute, get_rse_id
 from rucio.daemons.abacus import account as abacus_account
 from rucio.tests.common import execute, account_name_generator, rse_name_generator, file_generator, scope_name_generator
 from rucio.rse import rsemanager as rsemgr
@@ -1249,33 +1247,36 @@ class TestBinRucio():
 
     def test_list_account_usage(self):
         """ CLIENT (USER): list account usage. """
-        db_session = session.get_session()
-        db_session.query(models.AccountUsage).delete()
-        db_session.query(models.AccountLimit).delete()
-        db_session.query(models.AccountGlobalLimit).delete()
-        db_session.query(models.UpdatedAccountCounter).delete()
-        db_session.commit()
-        rse = 'MOCK4'
-        rse_id = get_rse_id(rse)
-        rse_exp = 'MOCK|MOCK4'
-        account = 'root'
-        usage = 4
-        local_limit = 10
-        local_left = local_limit - usage
-        global_limit = 20
-        global_left = global_limit - usage
-        self.account_client.set_local_account_limit(account, rse, local_limit)
-        self.account_client.set_global_account_limit(account, rse_exp, global_limit)
-        increase(rse_id, InternalAccount(account), 1, usage)
-        abacus_account.run(once=True)
-        cmd = 'rucio list-account-usage {0}'.format(account)
-        exitcode, out, err = execute(cmd)
-        print(out)
+        if environ.get('SUITE', 'all') != 'client':
+            from rucio.db.sqla import session, models
+            from rucio.core.account_counter import increase
+            db_session = session.get_session()
+            db_session.query(models.AccountUsage).delete()
+            db_session.query(models.AccountLimit).delete()
+            db_session.query(models.AccountGlobalLimit).delete()
+            db_session.query(models.UpdatedAccountCounter).delete()
+            db_session.commit()
+            rse = 'MOCK4'
+            rse_id = self.rse_client.get_rse(rse)['id']
+            rse_exp = 'MOCK|MOCK4'
+            account = 'root'
+            usage = 4
+            local_limit = 10
+            local_left = local_limit - usage
+            global_limit = 20
+            global_left = global_limit - usage
+            self.account_client.set_local_account_limit(account, rse, local_limit)
+            self.account_client.set_global_account_limit(account, rse_exp, global_limit)
+            increase(rse_id, InternalAccount(account), 1, usage)
+            abacus_account.run(once=True)
+            cmd = 'rucio list-account-usage {0}'.format(account)
+            exitcode, out, err = execute(cmd)
+            print(out)
 
-        nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse, usage, local_limit, local_left), out), None)
-        nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse_exp, usage, global_limit, global_left), out), None)
-        cmd = 'rucio list-account-usage --rse {0} {1}'.format(rse, account)
-        nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse, usage, local_limit, local_left), out), None)
-        nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse_exp, usage, global_limit, global_left), out), None)
-        self.account_client.set_local_account_limit(account, rse, -1)
-        self.account_client.set_global_account_limit(account, rse_exp, -1)
+            nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse, usage, local_limit, local_left), out), None)
+            nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse_exp, usage, global_limit, global_left), out), None)
+            cmd = 'rucio list-account-usage --rse {0} {1}'.format(rse, account)
+            nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse, usage, local_limit, local_left), out), None)
+            nose.tools.assert_not_equal(re.search('.*{0}.*{1}.*{2}.*{3}'.format(rse_exp, usage, global_limit, global_left), out), None)
+            self.account_client.set_local_account_limit(account, rse, -1)
+            self.account_client.set_global_account_limit(account, rse_exp, -1)
